Moved all the AI types into a single function.
Groovy, all the primitive types are done, with all the skill types and
equipment. Now I can create groups, and then once groups are done,
that's it for the Mission AI side of things. The fuller AI commander
will come as a part of v1.0, which will make AI do cleverer things based
on US Army Field Manuals for sniping and fire teams.