Completed a lot -- including crashes
Crashes are automated and support any vehicle in the game. Crashes are
dynamic and will appear anywhere. Admin functions aren't submitted yet
as I'm not working on client-side stuff until all the mission system is
completed.

The idea once this is done is to have dynamic missions, so that there is
no need for numbered files. The difficulties will be alterable, the
loot, the crash and of course, there will be more types of mission to
do.

I'm looking forward to introducing "capture" missions, where a non-bambi
player will be ambushed and their vehicle will then be seized and that
will become the mission point.

Also to be added, since the Sniper class is overachieving, sniper and
sniper + spotter missions where a sniper will take up a spot and kill
non-bambi players.