Humongous amount of code.
I've added a LOT, and the main body of code for the random mission
generation is complete. Missions are generated based on an array, and
arrays can be done randomly, too. The result of this is that if
individuals want to create specific missions (like I do), where AI
members are placed into specific buildings, and so on, then they can
just define the array, pop in the units and the mission is done.

This means no need for extra files like SM1 or what have you, just a
small array. The better the code is to handle this, the more sexy the
outcomes of AI missions.

At present -- not tested, so do not use. Release is VERY close, which is
exciting!