Some minor additions
Been busy this weekend, so not a lot of changes. However, tomorrow bodes
well for free time and conclusion of this development stage.

This is the prelude to loot boxes which can be made very easily, again,
1 or 2 lines, and use multiple types of item.

I know all you lovely people are desperate for the release, but be
patient. It's coming. I like good code quality, and you will appreciate
the end result more. :+1: